graphql_api/requirements.txt
graphene>=2.1,<3
graphene-django>=2.1,<3
graphql-core>=2.1,<3
django==3.1.14
django-filter>=2
pytest==7.4.3
pytest-django==4.7.0